src/clickup-automation/clickup_daily/requirements.txt
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
discord-interactions==0.4.0
discord-webhook==0.16.3
idna==3.3
loguru==0.6.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.20.0
requests==2.28.0
urllib3==1.26.9
